requirements.txt
aiogram==3.4.1
Jinja2==3.1.3
reportlab==4.1.0
pydantic==2.5.3
pydantic-settings==2.2.1
python-docx==1.1.0
